Fix glfwInit failures (#51)
For some reason glfwInit would fail on one of my Linux machines and if I
would make OpenNFS ignore the error, it would launch just fine. It turns
out glfw does not have good support for Wayland, so this commit makes it
run in X11 mode, which fixes the problem.